requirements.txt
face_recognition_models
Click>=6.0
numpy
Pillow
scipy>=0.17.0
ultralytics
face_recognition==1.3.0
dlib==19.22.0